SSR-3283: Refactored test helper code
