+++ __tests__/commands/authOptions.test.ts
@@ -1,11 +1,13 @@
-import path from 'path';
-
 import { program, CommanderStatic } from 'commander';
 import { afterAll, beforeAll, describe, expect, test } from '@jest/globals';
 
 import authCommand from '../../dist/commands/Auth/authCommand';
-import initCommand from '../../dist/commands/Initialize/initializeCommand';
-import { createTempDir, removeTempDir } from '../helpers';
+import {
+  createTempDir,
+  createTemplate,
+  initProject,
+  removeTempDir
+} from '../test-helpers';
 
 const PROJECT_NAME = 'test-project';
 const ORIGINAL_CWD = process.cwd();
@@ -34,8 +36,8 @@ describe('auth command: options', () => {
   beforeAll(async () => {
     createTempDir(TEMP_DIR);
 
-    await _initProject();
-    await _createTemplate();
+    await initProject(TEMP_DIR, PROJECT_NAME);
+    await createTemplate(TEMP_DIR, PROJECT_NAME);
 
     mainProgram = authCommand(program);
     process.argv = AUTH_OPTIONS;
@@ -74,37 +76,3 @@ describe('auth command: options', () => {
     expect(mainProgram.opts().proxy).toBe('http://localhost:8080');
   });
 });
-
-async function _initProject() {
-  process.chdir(path.join(TEMP_DIR));
-  global.currentDir = process.cwd();
-
-  initCommand(program);
-  process.argv = [
-    'node',
-    'dist',
-    'init',
-    '--quick',
-    '--project-name',
-    PROJECT_NAME
-  ];
-  await program.parseAsync(process.argv);
-}
-
-async function _createTemplate() {
-  process.chdir(TEMP_DIR + '/' + PROJECT_NAME);
-  global.currentDir = process.cwd();
-
-  initCommand(program);
-  process.argv = [
-    'node',
-    'dist',
-    'create-template',
-    '--quick',
-    '--app-name',
-    'test-app',
-    '--app-id',
-    '100'
-  ];
-  await program.parseAsync(process.argv);
-}
